requirements.txt
faiss_cpu==1.8.0
sentence-transformers==3.0.1
pandas==1.16.0